task-30-rag-app/pyproject.toml
[project]
name = "task-30-rag-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.12",
    "google-genai>=1.14.0",
    "numpy>=2.2.5",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.40",
    "uuid>=1.30",
    "uvicorn[standard]>=0.34.2",
]
